backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<2.0
django-cors-headers>=4.4,<4.5
django-storages>=1.14,<1.15
boto3>=1.35,<1.36
Pillow>=10.4,<10.5

